@@ -1,7 +1,5 @@
 package jlox.interpreters.lox;
 
-import java.io.IOException;
-
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeAll;
 
@@ -9,7 +7,7 @@ class BaseLoxTest {
 
     public static final String TEST_FILE_DIR = "../test/";
     @BeforeAll
-    public static void beforeClass() throws IOException {
+    public static void beforeClass() {
         Lox.clearState();
         LoxTestUtil.beforeClass();
     }
