@@ -9,20 +9,21 @@ public class InheritenceTest extends BaseLoxTest {
     
     @Test
     public void superMethodTest() {
-        final String loxCode = "class Doughnut {\r\n" + 
-                "  cook() {\r\n" + 
-                "    print \"Fry until golden brown.\";\r\n" + 
-                "  }\r\n" + 
-                "}\r\n" + 
-                "\r\n" + 
-                "class BostonCream < Doughnut {\r\n" + 
-                "  cook() {\r\n" + 
-                "    super.cook();\r\n" + 
-                "    print \"Pipe full of custard and coat with chocolate.\";\r\n" + 
-                "  }\r\n" + 
-                "}\r\n" + 
-                "\r\n" + 
-                "BostonCream().cook();";
+        final String loxCode = """
+                class Doughnut {\r
+                  cook() {\r
+                    print "Fry until golden brown.";\r
+                  }\r
+                }\r
+                \r
+                class BostonCream < Doughnut {\r
+                  cook() {\r
+                    super.cook();\r
+                    print "Pipe full of custard and coat with chocolate.";\r
+                  }\r
+                }\r
+                \r
+                BostonCream().cook();""";
         
         Lox.run(loxCode);
         
@@ -32,25 +33,26 @@ public class InheritenceTest extends BaseLoxTest {
     
     @Test
     public void superMethodFromSuperclassTest() {
-        final String loxCode = "class A {\r\n" + 
-                "  method() {\r\n" + 
-                "    print \"A method\";\r\n" + 
-                "  }\r\n" + 
-                "}\r\n" + 
-                "\r\n" + 
-                "class B < A {\r\n" + 
-                "  method() {\r\n" + 
-                "    print \"B method\";\r\n" + 
-                "  }\r\n" + 
-                "\r\n" + 
-                "  test() {\r\n" + 
-                "    super.method();\r\n" + 
-                "  }\r\n" + 
-                "}\r\n" + 
-                "\r\n" + 
-                "class C < B {}\r\n" + 
-                "\r\n" + 
-                "C().test();";
+        final String loxCode = """
+                class A {\r
+                  method() {\r
+                    print "A method";\r
+                  }\r
+                }\r
+                \r
+                class B < A {\r
+                  method() {\r
+                    print "B method";\r
+                  }\r
+                \r
+                  test() {\r
+                    super.method();\r
+                  }\r
+                }\r
+                \r
+                class C < B {}\r
+                \r
+                C().test();""";
         
         Lox.run(loxCode);
         
