@@ -7,21 +7,21 @@ import org.junit.jupiter.api.Test;
 class LogicTest extends BaseLoxTest {
 
     @Test
-    void trueAndtrue() {
+    void trueAndTrue() {
         Lox.run("print true and true;");
         LoxTestUtil.assertHasNoErrors();
         
         assertLineEquals("true");
     }
     @Test
-    void trueAndfalse() {
+    void trueAndFalse() {
         Lox.run("print true and false;");
         LoxTestUtil.assertHasNoErrors();
         
         assertLineEquals("false");
     }
     @Test
-    void falseAndfalse() {
+    void falseAndFalse() {
         Lox.run("print false and false;");
         LoxTestUtil.assertHasNoErrors();
         
@@ -29,21 +29,21 @@ class LogicTest extends BaseLoxTest {
     }
     
     @Test
-    void trueOrtrue() {
+    void trueOrTrue() {
         Lox.run("print true or true;");
         LoxTestUtil.assertHasNoErrors();
         
         assertLineEquals("true");
     }
     @Test
-    void trueOrfalse() {
+    void trueOrFalse() {
         Lox.run("print true or false;");
         LoxTestUtil.assertHasNoErrors();
         
         assertLineEquals("true");
     }
     @Test
-    void falseOrfalse() {
+    void falseOrFalse() {
         Lox.run("print false or false;");
         LoxTestUtil.assertHasNoErrors();
         
