@@ -1,15 +1,13 @@
 package jlox.interpreters.lox;
 
-import static jlox.interpreters.lox.LoxTestUtil.assertLineEquals;
-
-import java.io.IOException;
-
 import org.junit.jupiter.api.Test;
+
+import static jlox.interpreters.lox.LoxTestUtil.assertLineEquals;
 
 class PrintTest extends BaseLoxTest {
     
     @Test
-    void printString() throws IOException {
+    void printString() {
         Lox.run("print \"Hello World\";");
         
         LoxTestUtil.assertHasNoErrors();
@@ -17,26 +15,26 @@ class PrintTest extends BaseLoxTest {
         assertLineEquals("Hello World");
     }
     @Test
-    void printTrue() throws IOException {
+    void printTrue() {
         Lox.run("print true;");
         LoxTestUtil.assertHasNoErrors();
         assertLineEquals("true");
     }
     @Test
-    void printFalse() throws IOException {
+    void printFalse() {
         Lox.run("print false;");
         LoxTestUtil.assertHasNoErrors();
         assertLineEquals("false");
     }
     @Test
-    void printInteger() throws IOException {
+    void printInteger() {
         Lox.run("print 1;");
         LoxTestUtil.assertHasNoErrors();
         assertLineEquals("1");
 
     }
     @Test
-    void printFloat() throws IOException {
+    void printFloat() {
         Lox.run("print 1.1;");
         LoxTestUtil.assertHasNoErrors();
         assertLineEquals("1.1");
